Refactor resample step
The aim here is to be able to resample both the fixed and the moving image
of the register step.
Tho this end, the resampling was factorized in a widget so it can be used
twice without any code duplication. The step simply instantiates the
widgets.
The first resampling is done on the fixed image of the registration. The
second resampling is done on the moving image of the registration,
using the same final spacing as given by the first resampling.

Finaly, the dynamic property ignoreWorkflow was added. It allows widget to
stay hidden/visible even though their workflow property would suggest
otherwise. It is useful when ones want to hide part of a widget. Without it
changing the workflow level would most likely set it to visible. With it,
the workflow level is ignored and the widget stays hidden.