Automaticaly hide the backgrounf label
When using the PDF segmenter, the output segmentation as (usually) two
labels: the object and the background.
But when using it interactively, the background label can hide all the
image which isn't ideal for user interaction. That's why upo completion of
the PDF segmenter, the background label is set to 0 so only the segmented
mask is left.